requirements.txt
# Requirements for Langchain_Prompts project

streamlit  # For building web apps
# LangChain Core
langchain
langchain-core
# OpenAI Integration
langchain-openai
openai
# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

pydantic